CURB-4561 Renamed ReduxConnectedReactPortal to ConnectedReactPortal
